feat: modularize home view
